fix(terminal): a local OSC 7 claim is trusted only when it matches the kernel-reported shell cwd, since pty bytes carry no author and tty-ownership sampling races buffered parsing
